requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
opensearch-py==2.4.2
python-dotenv==1.0.1